requirements.txt
fastapi
uvicorn
faster-whisper
openai
streamlit
python-dotenv
pydub
